Refactor: Move BoostedNodesProvider to RealVoiceIO+Boosted
